+++ scripts/worker_port_warning.sh
@@ -1,4 +1,7 @@
 # Warn if non-standard worker_api_port is being used, likely default value they did not override properly
+if [[ "${blockchains}"  == "apple" && "${worker_api_port}" != '8947' ]]; then
+  echo "Apple worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8947?"
+fi
 if [[ "${blockchains}"  == "bpx" && "${worker_api_port}" != '8945' ]]; then
   echo "BPX worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8945?"
 fi
@@ -10,6 +13,9 @@ if [[ "${blockchains}"  == "cactus" && "${worker_api_port}" != '8936' ]]; then
 fi
 if [[ "${blockchains}"  == "chia" && "${worker_api_port}" != '8927' ]]; then
   echo "Chia worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8927?"
+fi
+if [[ "${blockchains}"  == "chinilla" && "${worker_api_port}" != '8948' ]]; then
+  echo "Chinilla worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8948?"
 fi
 if [[ "${blockchains}"  == "chives" && "${worker_api_port}" != '8931' ]]; then
   echo "Chives worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8931?"
@@ -26,14 +32,23 @@ fi
 if [[ "${blockchains}"  == "flora" && "${worker_api_port}" != '8932' ]]; then
   echo "Flora worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8932?"
 fi
+if [[ "${blockchains}"  == "gold" && "${worker_api_port}" != '8949' ]]; then
+  echo "Gold worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8949?"
+fi
 if [[ "${blockchains}"  == "hddcoin" && "${worker_api_port}" != '8930' ]]; then
   echo "HDDCoin worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8930?"
+fi
+if [[ "${blockchains}"  == "littlelambocoin" && "${worker_api_port}" != '8946' ]]; then
+  echo "LittleLamboCoin worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8946?"
 fi
 if [[ "${blockchains}"  == "maize" && "${worker_api_port}" != '8933' ]]; then
   echo "Maize worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8933?"
 fi
+if [[ "${blockchains}"  == "mint" && "${worker_api_port}" != '8950' ]]; then
+  echo "Mint worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8950?"
+fi
 if [[ "${blockchains}"  == "mmx" && "${worker_api_port}" != '8940' ]]; then
-  echo "Maize worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8940?"
+  echo "MMX worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8940?"
 fi
 if [[ "${blockchains}"  == "nchain" && "${worker_api_port}" != '8929' ]]; then
   echo "N-Chain worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8936?"
@@ -55,4 +70,10 @@ if [[ "${blockchains}"  == "staicoin" && "${worker_api_port}" != '8934' ]]; then
 fi
 if [[ "${blockchains}"  == "stor" && "${worker_api_port}" != '8935' ]]; then
   echo "Stor worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8935?"
+fi
+if [[ "${blockchains}"  == "tad" && "${worker_api_port}" != '8951' ]]; then
+  echo "Tad worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8951?"
+fi 
+if [[ "${blockchains}"  == "wheat" && "${worker_api_port}" != '8952' ]]; then
+  echo "Wheat worker with non-standard worker_api_port of ${worker_api_port} found.  Did you mean to use 8952?"
 fi